requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
opencv-python-headless==4.8.0.76
lime==0.2.0.1
matplotlib==3.7.1
gunicorn==21.2.0
scikit-image==0.21.0